shadowlands/requirements.txt
git+https://github.com/kayagoban/asciimatics
git+https://github.com/kayagoban/python-qrcode.git
git+https://github.com/kayagoban/schedule.git
asn1crypto==0.24.0
attrdict==2.0.0
bitcoin==1.1.42
certifi==2018.8.24
cffi==1.12.2
chardet==3.0.4
Click==7.0
coincurve==11.0.0
construct==2.9.45
cryptocompy==0.1.1.dev1
#cytoolz==0.9.0.1
ecdsa==0.13.3
ECPy==0.9.1
future==0.16.0
hexbytes==0.1.0
hidapi
hurry.filesize==0.9
idna==2.7
ledgerblue
libusb1==1.6.6
lru-dict==1.1.6
mnemonic==0.18
mypy-extensions==0.4.1
numpy==1.15.2
parsimonious==0.8.0
pathlib==1.0.1
pbkdf2==1.3
py-ecc==1.6.0
py-solc==3.1.0
pyblake2==1.1.2
pycparser==2.19
pycryptodome==3.6.6
pycryptodomex==3.6.6
pyethash==0.1.27
pyfiglet==0.7.5
pyperclip==1.7.0
pysha3==1.0.2
python-u2flib-host==3.0.3
python3-wget==0.0.2b1
PyYAML==5.1
repoze.lru==0.7
scrypt==0.8.6
semantic-version==2.6.0
six==1.11.0
toolz==0.9.0
trezor==0.11.2
urllib3==1.24.2
wcwidth==0.1.7
web3==5.9.0
wget==3.2
cached_property==1.5.1
